Refactor InfrahubEventFilter
